requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
autopep8==1.4.4
backcall==0.1.0
bleach==3.1.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cloudpickle==1.1.1
cryptography==2.7
cycler==0.10.0
daemonize==2.5.0
dask==1.2.2
decorator==4.4.0
defusedxml==0.6.0
docutils==0.14
entrypoints==0.3
future==0.17.1
gast==0.2.2
google-pasta==0.1.6
graphviz==0.11
grpcio==1.20.1
guildai==0.6.4
h5py==2.9.0
idna==2.8
imageio==2.5.0
ipdb==0.12
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.1
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
pkginfo==1.5.0.1
pluggy==0.11.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==3.0.0
pydot==1.4.1
pydotplus==2.0.2
pyflakes==2.1.1
Pygments==2.4.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
python-jsonrpc-server==0.1.2
python-language-server==0.26.1
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.1
qtconsole==4.4.4
readme-renderer==24.0
requests==2.21.0
requests-toolbelt==0.9.1
rope==0.14.0
scikit-image==0.14.2
scikit-learn==0.20.2
scikit-optimize==0.5.2
scipy==1.2.0
Send2Trash==1.5.0
six==1.12.0
snowballstemmer==1.2.1
tabview==1.4.3
tb-nightly==1.14.0a20190603
tensorboard==1.14.0
tensorboardX==1.7
tensorflow-estimator==1.13.0
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019060501
tfp-nightly==0.7.0.dev20190524
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.1
tornado==6.0.2
tqdm==4.31.0
traitlets==4.3.2
twine==1.13.0
typed-ast==1.3.5
urllib3==1.24.3
virtualenv==16.6.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
Whoosh==2.7.4
widgetsnbextension==3.4.2
wrapt==1.11.1
yapf==0.27.0
